+++ backend/canisters/group/api/src/updates/mod.rs
@@ -15,7 +15,6 @@ pub mod c2c_invite_users;
 pub mod c2c_join_group;
 pub mod c2c_leave_group;
 pub mod c2c_local_index;
-pub mod c2c_report_message_v2;
 pub mod c2c_send_message;
 pub mod c2c_set_user_suspended;
 pub mod c2c_start_import_into_community;