+++ backend/canisters/group/impl/src/updates/mod.rs
@@ -12,7 +12,6 @@ pub mod c2c_join_group;
 pub mod c2c_leave_group;
 pub mod c2c_local_index;
 pub mod c2c_notify_p2p_swap_status_change;
-pub mod c2c_report_message_v2;
 pub mod c2c_set_user_suspended;
 pub mod c2c_start_import_into_community;
 pub mod c2c_tip_message;