+++ backend/canisters/local_user_index/api/src/main.rs
@@ -33,7 +33,6 @@ fn main() {
     generate_ts_method!(local_user_index, join_community);
     generate_ts_method!(local_user_index, join_group);
     generate_ts_method!(local_user_index, register_user);
-    generate_ts_method!(local_user_index, report_message_v2);
     generate_ts_method!(local_user_index, uninstall_bot);
     generate_ts_method!(local_user_index, withdraw_from_icpswap);
 }