+++ backend/integration_tests/src/client/local_user_index.rs
@@ -24,7 +24,6 @@ generate_msgpack_update_call!(join_channel);
 generate_msgpack_update_call!(join_community);
 generate_msgpack_update_call!(join_group);
 generate_msgpack_update_call!(register_user);
-generate_msgpack_update_call!(report_message_v2);
 generate_msgpack_update_call!(uninstall_bot);
 
 pub mod happy_path {